Changed Architecture,

+++ DigitalLibrary/DigitalLibrary.Web/Controllers/TrustedUserWorkController.cs
@@ -4,37 +4,47 @@
     using System.Web.Mvc;
 
     using Kendo.Mvc.UI;
+    using Kendo.Mvc.Extensions;
 
     using DigitalLibrary.Data;
-    using DigitalLibrary.Logic;
     using DigitalLibrary.Web.Models;
+    using DigitalLibrary.Web.ViewModels.Work;
+    using DigitalLibrary.Data.Logic;
 
-    [Authorize(Roles = "trusted")]
+    // [Authorize(Roles = "trusted")]
     public class TrustedUserWorkController : BaseController
     {
         private const int TrustedRoleNeededRating = 70;
         private const int MinimumPositiveUploadsToBecomeTrusted = 3;
-       
-        public TrustedUserWorkController(ILibraryData data)
+
+        public TrustedUserWorkController(IDigitalLibraryData data)
             : base(data)
         {
         }
 
-        public ActionResult List([DataSourceRequest]DataSourceRequest request)
+        public ActionResult List()
+        {
+            return this.View();
+        }
+
+        [HttpPost]
+        public ActionResult Read([DataSourceRequest]DataSourceRequest request)
         {
             if (request.PageSize == 0)
             {
                 request.PageSize = 5;
             }
 
-            var data = this.GetAllUnApprovedWorks();
+            var data = this.GetAllUnApprovedWorks()
+                .ToDataSourceResult(request);
 
-            return this.View(data);
+            return this.Json(data);
+            //return this.View(data);
         }
 
         public ActionResult Approove(int id)
         {
-            var workToBeApprooved = this.Data.Works.Find(id);
+            var workToBeApprooved = this.Data.Works.GetById(id);
             var uploadedBy = workToBeApprooved.UploadedBy;
 
             uploadedBy.PositiveUploads++;
@@ -46,43 +56,45 @@
             {
                 this.IdentityManager.AddUserToRole(uploadedBy.Id, "trusted");
             }
-
+            Response.Redirect("~/TrustedUserWork/List");
             return this.View("List");
         }
 
-        public ActionResult Delete(int id)
+        public ActionResult Delete([DataSourceRequest]DataSourceRequest request, WorkPublicListViewModel model)
         {
-            var workToBeDeleted = this.Data.Works.Find(id);
-
-            var uploadedBy = workToBeDeleted.UploadedBy;
-            uploadedBy.NegativeUploads++;
-           
-            this.Data.Works.Delete(workToBeDeleted);
-
-            var folders = workToBeDeleted.ZipFileLink.Split('/').ToList();
-
-            folders.RemoveAt(folders.Count - 1);
-
-            var filePath = string.Join("/", folders);
-
-            FileManager.DeleteFile(filePath);
-
-            this.Data.SaveChanges();
-
-            if (uploadedBy.Rating < TrustedRoleNeededRating)
+            if (model != null && ModelState.IsValid)
             {
-                this.IdentityManager.ClearUserRoles(uploadedBy.Id, "trusted");
+                var workToBeDeleted = this.Data.Works.GetById(model.Id);
+
+                var uploadedBy = workToBeDeleted.UploadedBy;
+                uploadedBy.NegativeUploads++;
+
+                this.Data.Works.Delete(workToBeDeleted);
+                this.Data.SaveChanges();
+
+                var folders = workToBeDeleted.ZipFileLink.Split('/').ToList();
+
+                folders.RemoveAt(folders.Count - 1);
+
+                var filePath = string.Join("/", folders);
+
+                FileManager.DeleteFile(filePath);
+
+                if (uploadedBy.Rating < TrustedRoleNeededRating)
+                {
+                    this.IdentityManager.ClearUserRoles(uploadedBy.Id, "trusted");
+                }
             }
 
-            return this.View("List");
+            return this.Json(new[] { model }.ToDataSourceResult(request, ModelState));
         }
 
-        private IQueryable<WorkListViewModel> GetAllUnApprovedWorks()
+        private IQueryable<WorkPublicListViewModel> GetAllUnApprovedWorks()
         {
             var unApprovedWorks = this.Data.Works
                 .All()
                 .Where(w => !w.IsApproved)
-                .Select(WorkListViewModel.FromWork);
+                .Select(WorkPublicListViewModel.FromWork);
 
             return unApprovedWorks;
         }
